[#17][ADD] Added Object management instead of assets type management. Page now regroups IOC type and assets. More to come
