+++ source/app/util.py
@@ -112,7 +112,8 @@ class AlchemyEncoder(json.JSONEncoder):
 
             # an SQLAlchemy class
             fields = {}
-            for field in [x for x in dir(obj) if not x.startswith('_') and x != 'metadata']:
+            for field in [x for x in dir(obj) if not x.startswith('_') and x != 'metadata'
+                                                 and x != 'query' and x != 'query_class']:
                 data = obj.__getattribute__(field)
                 try:
                     json.dumps(data)  # this will fail on non-encodable values, like other classes
